+++ core-ng/src/test/java/core/framework/internal/db/RepositoryImplAutoIncrementIdEntityTest.java
@@ -34,7 +34,7 @@ class RepositoryImplAutoIncrementIdEntityTest {
     void createDatabase() {
         database = new DatabaseImpl("db");
         database.url("jdbc:hsqldb:mem:mysql;sql.syntax_mys=true");
-        database.execute("CREATE TABLE auto_increment_id_entity (id INT AUTO_INCREMENT PRIMARY KEY, string_field VARCHAR(20), double_field DOUBLE, enum_field VARCHAR(10), date_time_field TIMESTAMP, zoned_date_time_field TIMESTAMP)");
+        database.execute("CREATE TABLE auto_increment_id_entity (id INT AUTO_INCREMENT PRIMARY KEY, string_field VARCHAR(20), double_field DOUBLE, enum_field VARCHAR(10), date_time_field DATETIME, zoned_date_time_field TIMESTAMP)");
         repository = database.repository(AutoIncrementIdEntity.class);
     }